+++ src/types/index.ts
@@ -1,8 +1,0 @@
-export interface CountryCode {
-  id?: string;
-  code: number | string;
-  name: string;
-  abbreviation: string;
-}
-
-export * from './iceIconsEnum';