+++ vite.config.ts
@@ -2,7 +2,6 @@ import { resolve } from 'node:path';
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
-import { iceAliases } from './vite.aliases';
 
 export default defineConfig({
   plugins: [
@@ -10,11 +9,10 @@ export default defineConfig({
     dts({
       include: ['src'],
       outDir: 'dist',
-      rollupTypes: false,
+      rollupTypes: true,
       tsconfigPath: './tsconfig.lib.json',
     }),
   ],
-  resolve: { alias: iceAliases },
   build: {
     lib: {
       entry: resolve(__dirname, 'src/index.ts'),
@@ -23,26 +21,7 @@ export default defineConfig({
       fileName: (format) => (format === 'es' ? 'index.js' : 'index.cjs'),
     },
     rollupOptions: {
-      external: [
-        'react',
-        'react-dom',
-        'react/jsx-runtime',
-        /^@mui\//,
-        /^@emotion\//,
-        /^@fontsource\//,
-        'i18next',
-        'react-i18next',
-        'moment',
-        'date-fns',
-        'classnames',
-        'react-hook-form',
-        'react-datepicker',
-        'react-icons',
-        'react-country-flag',
-        'react-use',
-        'stylis',
-        'stylis-plugin-rtl',
-      ],
+      external: ['react', 'react-dom', 'react/jsx-runtime', 'date-fns', 'moment', 'react-datepicker'],
     },
     cssCodeSplit: false,
     emptyOutDir: true,